FIX blend render with new logic - remove sepate "blend_oit" flag   * assume "blend + gbuffer" handles it

+++ src/generator/NodeGenerator.cpp
@@ -128,7 +128,7 @@ void NodeGenerator::bindBatch(
     const RenderContext& ctx,
     mesh::MeshType* type,
     const std::function<Program* (const mesh::LodMesh&)>& programSelector,
-    unsigned int kindBits,
+    uint8_t kindBits,
     render::Batch& batch,
     Node& container)
 {
