+++ src/generator/TextGenerator.cpp
@@ -126,7 +126,7 @@ void TextGenerator::bindBatch(
     const RenderContext& ctx,
     mesh::MeshType* type,
     const std::function<Program* (const mesh::LodMesh&)>& programSelector,
-    unsigned int kindBits,
+    uint8_t kindBits,
     render::Batch& batch,
     Node& container)
 {
