+++ src/generator/TextGenerator.h
@@ -51,7 +51,7 @@ public:
         const RenderContext& ctx,
         mesh::MeshType* type,
         const std::function<Program* (const mesh::LodMesh&)>& programSelector,
-        unsigned int kindBits,
+        uint8_t kindBits,
         render::Batch& batch,
         Node& container) override;
