+++ src/model/Node.h
@@ -86,7 +86,7 @@ public:
         const RenderContext& ctx,
         mesh::MeshType* type,
         const std::function<Program* (const mesh::LodMesh&)>& programSelector,
-        unsigned int kindBits,
+        uint8_t kindBits,
         render::Batch& batch) noexcept;
 
     inline Node* getParent() const noexcept {
